requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.0

# Database
psycopg2-binary==2.9.9

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client for Inter-service Communication
requests==2.31.0
urllib3==2.1.0

# CORS Support
Flask-CORS==4.0.0

# Testing (optional but recommended for assignment)
pytest==7.4.3
pytest-cov==4.1.0

# Production Server (optional)
gunicorn==21.2.0